requirements.txt
requests
beautifulsoup4
lxml
python-dotenv
pandas
ipykernel
google-cloud-bigquery